refactor: 패키지 구조 변경
질문 게시판과 회원관리 게시판을 구분해서 담당하도록 패키지 구조를 변경한다.
