+++ src/main/java/com/codessquad/domain/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.codessquad.domain;
+package com.codessquad.domain.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
